[BE] refactor: change the type of battingAverage
- from Long to Double

+++ BE/src/main/java/com/codesquad/baseball06/model/Player.java
@@ -6,16 +6,16 @@ public class Player {
   private Long id;
   private String name;
   private PlayerType type;
-  private Long battingAverage;
+  private Double battingAverage;
 
-  private Player(Long id, String name, PlayerType type, Long battingAverage) {
+  private Player(Long id, String name, PlayerType type, Double battingAverage) {
     this.id = id;
     this.name = name;
     this.type = type;
     this.battingAverage = battingAverage;
   }
 
-  private Player(String name, PlayerType type, Long battingAverage) {
+  private Player(String name, PlayerType type, Double battingAverage) {
     this.name = name;
     this.type = type;
     this.battingAverage = battingAverage;
@@ -33,15 +33,19 @@ public class Player {
     return type;
   }
 
-  public Long getBattingAverage() {
+  public Double getBattingAverage() {
     return battingAverage;
   }
 
-  public static Player create(Long id, String name, PlayerType type, Long battingAverage) {
+  public static Player create(Long id, String name, PlayerType type, Double battingAverage) {
     return new Player(id, name, type, battingAverage);
   }
 
-  public static Player create(String name, PlayerType type, Long battingAverage) {
+  public static Player create(String name, PlayerType type, Double battingAverage) {
     return new Player(name, type, battingAverage);
   }
+
+
+
+
 }
